Add moment.js and lowdash
Harp's biggest flaw is how it handles exernal libraies. It's a racket.
The key is to compile the js file to one line and include it in the jade
file. The file output doesn't actually include any of the JS file. It is
just used to compile the `.jada` template into valide `.html`. Same if
you are using it served.

We could put thought some thought on how to create a gulp tosk.

For now, each library is included in scipts. A possible optimization
would be to have one separate file for each library and include those in
the `_scripts.master.jade` file.